requirements.txt
Flask
mysql-connector-python
flask-cors
python-dotenv
gunicorn